.github: run CI on staging branches, too
The GitHub Actions workflows from upstream limit push triggers to just
hte master and release branches. Trouble is, the Mirantis/ee-engine
repo has branch protection rules which block force-pushes on release
branches, which makes it awkward to perform a CI run with the
patchset-oriented development paradigm. Enable CI on "staging" branches
which are not protected and can be freely force-pushed to.

Signed-off-by: Cory Snider <csnider@mirantis.com>