overlay: Add State.IPAM support in Network inspection
Adds support for inspecting overlay networks with a State object
that includes IPAM state and IP utilization details.

To enable this, the cnmallocator is extended with a new
UpdateNetworkState() method, used in a callback from the Swarm API
server to populate IPAM data in the network State object. This
State object is passed to the API server as a blob, avoiding
further changes to the Swarm API and preserving backward
compatibility in future releases.

Signed-off-by: Andrey Epifanov <aepifanov@mirantis.com>

# Conflicts:
#	go.mod